Remove Releases tab from header
This pull request moves the content of the Releases tab to the version number next to the title. With the search bar, the header was getting too crowded.

I've cleaned up the search style a bit and made it look like the React one (with the background color).
I've also improved the styling of the versions page